check species without a taxon key in geodata request
Move function species_without_speciesKey() from image_request.py to data/common_api_request.py
and call it from both geodata_requets.py and image_request.py.
